requirements.txt
# Core dependencies for LocalWhisperTyping
openai-whisper>=20231117
sounddevice>=0.4.6
scipy>=1.9.0
numpy>=1.21.0
keyboard>=0.13.5
pystray>=0.19.4
Pillow>=9.0.0

# Additional dependencies that may be required
torch>=1.13.0
torchaudio>=0.13.0
